# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Anomaly Detection API",
    "description": "Backend foundation for Machine Learning Enhanced Graph-Based Anomaly Detection for Financial Transactions. Provides transaction ingestion, graph construction, and behavioral/temporal feature extraction.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/graph/{dataset_id}/summary": {
      "get": {
        "tags": [
          "Graph Analysis"
        ],
        "summary": "Transaction Graph Network Summary",
        "description": "Returns topological network statistics calculated using NetworkX for the specified dataset, including node count, directed edge count, graph density, connected components, and highest-degree hub nodes.",
        "operationId": "get_graph_summary_api_v1_graph__dataset_id__summary_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/{dataset_id}/nodes/{user_id}": {
      "get": {
        "tags": [
          "Graph Analysis"
        ],
        "summary": "Node Structural Metrics",
        "description": "Retrieve structural graph metrics (in/out degree, weighted volume, betweenness centrality) for a specific entity within a dataset.",
        "operationId": "get_node_metrics_api_v1_graph__dataset_id__nodes__user_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeMetrics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluation/{dataset_id}/compare": {
      "get": {
        "tags": [
          "Research Evaluation & Comparison"
        ],
        "summary": "Compare All Trained Experiments",
        "description": "Evaluates all trained experiments for the given dataset against stored ground-truth labels. Returns a comparison table ranked by PR-AUC (Average Precision) descending. Captures failed experiments explicitly with status='error' rather than silently omitting them. Clearly groups and distinguishes results by evaluation_mode (in-sample vs held-out). Highlights PR-AUC as the headline metric due to extreme positive-class imbalance in financial fraud.",
        "operationId": "compare_dataset_experiments_api_v1_evaluation__dataset_id__compare_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "include_curves",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to include full ROC and Precision-Recall curve point arrays for figure plotting",
              "default": false,
              "title": "Include Curves"
            },
            "description": "Whether to include full ROC and Precision-Recall curve point arrays for figure plotting"
          },
          {
            "name": "split_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional dataset split label to evaluate experiments in held-out mode against test partition",
              "title": "Split Label"
            },
            "description": "Optional dataset split label to evaluate experiments in held-out mode against test partition"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluation/{dataset_id}/export": {
      "post": {
        "tags": [
          "Research Evaluation & Comparison"
        ],
        "summary": "Export Research Run Results",
        "description": "Exports comprehensive evaluation comparison results, experiment manifests, and metric tables into persistent JSON and CSV artifacts under data/results/.",
        "operationId": "export_research_run_results_api_v1_evaluation__dataset_id__export_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluation/{dataset_id}/{experiment_label}": {
      "get": {
        "tags": [
          "Research Evaluation & Comparison"
        ],
        "summary": "Evaluate Single Experiment",
        "description": "Computes detailed classification (confusion matrix, precision, recall, F1, accuracy) and ranking (ROC-AUC, PR-AUC) metrics for a single trained experiment against ground-truth labels. If the experiment was trained with a split or split_label is provided, evaluates against the held-out test partition.",
        "operationId": "evaluate_single_experiment_api_v1_evaluation__dataset_id___experiment_label__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "experiment_label",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Label"
            }
          },
          {
            "name": "include_curves",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to include full ROC and Precision-Recall curve point arrays for figure plotting",
              "default": false,
              "title": "Include Curves"
            },
            "description": "Whether to include full ROC and Precision-Recall curve point arrays for figure plotting"
          },
          {
            "name": "split_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional split label to override or specify held-out test partition evaluation",
              "title": "Split Label"
            },
            "description": "Optional split label to override or specify held-out test partition evaluation"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentEvaluationMetrics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DatasetLabelsSummaryResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "UUID of the dataset"
          },
          "total_labels_uploaded": {
            "type": "integer",
            "title": "Total Labels Uploaded",
            "description": "Total rows in the uploaded label file"
          },
          "matched_count": {
            "type": "integer",
            "title": "Matched Count",
            "description": "Number of labels matched to existing dataset entities"
          },
          "unmatched_count": {
            "type": "integer",
            "title": "Unmatched Count",
            "description": "Number of label rows rejected because entity ID was not found"
          },
          "positive_count": {
            "type": "integer",
            "title": "Positive Count",
            "description": "Number of positive (anomalous/fraud) labels"
          },
          "negative_count": {
            "type": "integer",
            "title": "Negative Count",
            "description": "Number of negative (normal) labels"
          },
          "prevalence_rate": {
            "type": "number",
            "title": "Prevalence Rate",
            "description": "Prevalence rate of positive labels among matched entities (positives / matched)"
          },
          "rejected_errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rejected Errors",
            "description": "Detailed rejection messages for unmatched or unparseable label rows"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "total_labels_uploaded",
          "matched_count",
          "unmatched_count",
          "positive_count",
          "negative_count",
          "prevalence_rate"
        ],
        "title": "DatasetLabelsSummaryResponse",
        "description": "Summary of ground-truth labels attached to a dataset."
      },
...
      "EvaluationComparisonResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "UUID of the evaluated dataset"
          },
          "total_experiments_evaluated": {
            "type": "integer",
            "title": "Total Experiments Evaluated",
            "description": "Total count of experiments attempted in this run"
          },
          "successful_experiments_count": {
            "type": "integer",
            "title": "Successful Experiments Count",
            "description": "Count of successfully evaluated experiments",
            "default": 0
          },
          "failed_experiments_count": {
            "type": "integer",
            "title": "Failed Experiments Count",
            "description": "Count of experiments that encountered an evaluation error",
            "default": 0
          },
          "labels_summary": {
            "$ref": "#/components/schemas/DatasetLabelsSummaryResponse",
            "description": "Summary of ground-truth labels used"
          },
          "evaluation_mode": {
            "type": "string",
            "title": "Evaluation Mode",
            "description": "Protocol mode: 'in_sample' (pipeline validation), 'held_out' (split test set), or 'mixed'",
            "default": "in_sample"
          },
          "split_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Label",
            "description": "Optional split label if evaluation was restricted to a specific split"
          },
          "evaluation_mode_note": {
            "type": "string",
            "title": "Evaluation Mode Note",
            "description": "Notice on in-sample vs held-out evaluation validity",
            "default": "In-sample evaluation represents pipeline validation. For publication, evaluations must be performed on a held-out test split or time-based partition."
          },
          "by_evaluation_mode": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/ExperimentEvaluationMetrics"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "By Evaluation Mode",
            "description": "Experiments explicitly partitioned by evaluation_mode ('in_sample' vs 'held_out') to prevent conflation"
          },
          "headline_metric": {
            "type": "string",
            "title": "Headline Metric",
            "description": "Primary ranking metric",
            "default": "pr_auc"
          },
          "class_imbalance_note": {
            "type": "string",
            "title": "Class Imbalance Note",
            "default": "In financial fraud detection, the positive (anomalous) class is rare (<5%). Under severe class imbalance, accuracy can be misleadingly high. PR-AUC (Average Precision) is the primary headline metric for ranking configurations."
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentEvaluationMetrics"
            },
            "type": "array",
            "title": "Experiments",
            "description": "List of experiment evaluation metrics sorted by PR-AUC descending (successful first, followed by errors)"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "total_experiments_evaluated",
          "labels_summary",
          "experiments"
        ],
        "title": "EvaluationComparisonResponse",
        "description": "Multi-experiment performance comparison table sorted by PR-AUC descending."
      },
...
      "GraphSummaryResponse": {
        "properties": {
          "nodes": {
            "type": "integer",
            "title": "Nodes",
            "description": "Total number of unique user/account nodes"
          },
          "edges": {
            "type": "integer",
            "title": "Edges",
            "description": "Total number of directed transaction edges"
          },
          "density": {
            "type": "number",
            "title": "Density",
            "description": "Network density (actual edges / possible edges)"
          },
          "is_directed": {
            "type": "boolean",
            "title": "Is Directed",
            "description": "Whether the graph is directed",
            "default": true
          },
          "weakly_connected_components": {
            "type": "integer",
            "title": "Weakly Connected Components",
            "description": "Count of weakly connected subgraphs"
          },
          "strongly_connected_components": {
            "type": "integer",
            "title": "Strongly Connected Components",
            "description": "Count of strongly connected subgraphs"
          },
          "top_in_degree_nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeMetrics"
            },
            "type": "array",
            "title": "Top In Degree Nodes",
            "description": "Top nodes with highest incoming transaction connections"
          },
          "top_out_degree_nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeMetrics"
            },
            "type": "array",
            "title": "Top Out Degree Nodes",
            "description": "Top nodes with highest outgoing transaction connections"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "density",
          "weakly_connected_components",
          "strongly_connected_components"
        ],
        "title": "GraphSummaryResponse",
        "description": "Statistical summary of the constructed transaction graph."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeMetrics": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "in_degree": {
            "type": "integer",
            "title": "In Degree"
          },
          "out_degree": {
            "type": "integer",
            "title": "Out Degree"
          },
          "total_degree": {
            "type": "integer",
            "title": "Total Degree"
          },
          "weighted_in_degree": {
            "type": "number",
            "title": "Weighted In Degree"
          },
          "weighted_out_degree": {
            "type": "number",
            "title": "Weighted Out Degree"
          },
          "betweenness_centrality": {
            "type": "number",
            "title": "Betweenness Centrality"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "in_degree",
          "out_degree",
          "total_degree",
          "weighted_in_degree",
          "weighted_out_degree",
          "betweenness_centrality"
        ],
        "title": "NodeMetrics"
      },
...
      "ResearchExportResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "UUID of the exported dataset"
          },
          "export_timestamp": {
            "type": "string",
            "title": "Export Timestamp",
            "description": "UTC timestamp of the export"
          },
          "json_path": {
            "type": "string",
            "title": "Json Path",
            "description": "File path of exported JSON results"
          },
          "csv_path": {
            "type": "string",
            "title": "Csv Path",
            "description": "File path of exported CSV results table"
          },
          "total_experiments": {
            "type": "integer",
            "title": "Total Experiments",
            "description": "Count of experiments included in the export"
          },
          "successful_experiments": {
            "type": "integer",
            "title": "Successful Experiments",
            "description": "Count of successfully evaluated experiments"
          },
          "failed_experiments": {
            "type": "integer",
            "title": "Failed Experiments",
            "description": "Count of failed experiments"
          },
          "evaluation_mode": {
            "type": "string",
            "title": "Evaluation Mode",
            "description": "Evaluation mode (e.g. 'in_sample')"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "export_timestamp",
          "json_path",
          "csv_path",
          "total_experiments",
          "successful_experiments",
          "failed_experiments",
          "evaluation_mode"
        ],
        "title": "ResearchExportResponse",
        "description": "Response confirming research run export to JSON and CSV artifacts."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}